refactor(#3479): shrink tmux_runtime.rs below giant threshold
Behavior-preserving decomposition of the registered giant
src/services/discord/turn_bridge/tmux_runtime.rs (1545 prod LoC) below
the 1000 prod-LoC threshold. Three cohesive, verbatim clusters move into
tmux_runtime/ child modules; the async interrupt/cancel/hard-stop
orchestration + session-teardown logic stays in the module root.

- tmux_runtime/interrupt_policy.rs (225 prod): pure provider
  turn-interrupt policy decisions + value types (ProviderTurnInterruptPlan,
  ProviderTurnInterruptOutcome, the #3029(A) missed-SIGINT-target detection,
  the #3021 fallback gating, the #3207 claude session-preserving delivery
  selection + control envelope, and the #3169 teardown-SIGINT sentinel +
  suppression decision) plus their three pure unit-test modules.
- tmux_runtime/process_table.rs (248 prod): ps-backed process-table
  discovery (ProcessRow, provider_cli_pid_in_tmux, the descendant-walk +
  command-match scoring, pane-foreground classification, the wrapper-FIFO
  writer, and the send_sigint primitive).
- tmux_runtime/pid_exit.rs (176 prod): the #2426 OS-level PID-exit
  observation family (kqueue NOTE_EXIT / pidfd_open+poll / bounded-sleep
  fallback) plus its tests.

Bodies are byte-identical; only visibility prefixes, module paths, and
rustfmt wrapping changed. tmux_runtime.rs: 1545 -> 964 prod LoC; removed
from giant_file_registry.toml (grandfathered + grandfathered_baseline_paths)
and the giant ratchet baseline. ARCHITECTURE.md regenerated; change-surfaces
updated with the graduation + the three new submodule entries.
