+++ ARCHITECTURE.md
@@ -541,6 +541,10 @@ src/
 │   │   │   ├── completion_guard/
 │   │   │   │   ├── completion_context.rs
 │   │   │   │   └── completion_postgres.rs
+│   │   │   ├── tmux_runtime/
+│   │   │   │   ├── interrupt_policy.rs
+│   │   │   │   ├── pid_exit.rs
+│   │   │   │   └── process_table.rs
 │   │   │   ├── cancel_finalize_policy.rs
 │   │   │   ├── completion_guard.rs
 │   │   │   ├── context_window.rs
