+++ src/services/discord/turn_bridge/tmux_runtime.rs
@@ -5,6 +5,28 @@ use crate::services::provider::{CancelToken, ProviderKind};
 use crate::services::tmux_diagnostics::record_tmux_exit_reason;
 use std::time::Duration;
 
+// #3479: behavior-preserving decomposition of this giant module. The pure
+// interrupt-policy decisions, the `ps`-backed process-table discovery, and the
+// OS-level PID-exit observation moved verbatim into sibling leaf modules; the
+// async orchestration + session-teardown logic stays here and reaches the
+// moved items by their original bare names via these glob/explicit re-imports.
+mod interrupt_policy;
+mod pid_exit;
+mod process_table;
+
+use interrupt_policy::*;
+use pid_exit::wait_for_pid_exit;
+use process_table::{
+    pane_foreground_is_provider_wrapper, provider_cli_pid_in_tmux, send_sigint,
+    write_line_to_wrapper_fifo,
+};
+
+// #3169: `mod.rs`'s cancel epilogue records this sentinel via the
+// `tmux_runtime::ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON` path, so re-export it
+// from the module root at its original visibility now that it lives in the
+// `interrupt_policy` leaf module.
+pub(in crate::services::discord) use interrupt_policy::ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON;
+
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub(crate) enum TmuxCleanupPolicy {
     PreserveSession,
@@ -49,201 +71,6 @@ const PROVIDER_IDLE_CONFIRM_DELAY: Duration = Duration::from_millis(400);
 /// SIGKILL. Same #2426 rationale as `PROVIDER_INTERRUPT_SETTLE`: when the
 /// provider exits cleanly we observe its PID exit and proceed immediately.
 const PROVIDER_HARD_STOP_GRACE: Duration = Duration::from_millis(1500);
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-struct ProviderTurnInterruptPlan {
-    keys: &'static [&'static str],
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(in crate::services::discord) struct ProviderTurnInterruptOutcome {
-    pub tmux_session: Option<String>,
-    pub sent_keys: bool,
-    pub fallback_sigint_pid: Option<u32>,
-    pub missing_tmux_session: bool,
-    /// #3029(A): set when the SIGINT-only interrupt path (empty key list —
-    /// claude, whose pane C-c targets the wrapper) needed to deliver a SIGINT
-    /// to an actively-generating turn but the provider PID lookup returned
-    /// `None` (ps failure, command-name drift, or a just-spawned child not yet
-    /// visible). On that path the interrupt is the *only* delivery mechanism,
-    /// so a `None` PID is a silent no-op: the mailbox marks the turn [Stopped]
-    /// but no signal reaches the provider. This flag converts that into an
-    /// explicit failure the hard-stop path can escalate on, instead of
-    /// reporting unconditional success.
-    pub sigint_target_missing: bool,
-}
-
-/// #3029(A): does this provider/plan combination treat the direct SIGINT
-/// fallback as its *only* interrupt delivery (i.e. there is no send-keys path
-/// that could have reached the turn)? Claude uses an empty key list because a
-/// pane C-c hits the wrapper and tears the session down (#1260) — so when its
-/// SIGINT target is missing, the interrupt silently did nothing.
-fn interrupt_is_sigint_only(provider: &ProviderKind, plan_keys_empty: bool) -> bool {
-    plan_keys_empty && matches!(provider, ProviderKind::Claude)
-}
-
-/// #3029(A): the interrupt silently did nothing when the SIGINT-only path was
-/// the only delivery mechanism, the turn was genuinely active
-/// (`ready_for_input == false`), yet no SIGINT target PID could be resolved.
-/// An idle pane (`ready_for_input == true`) intentionally resolves to no PID
-/// and is NOT a missed interrupt (#3021).
-fn interrupt_sigint_target_missing(
-    provider: &ProviderKind,
-    plan_keys_empty: bool,
-    ready_for_input: bool,
-    resolved_sigint_pid: Option<u32>,
-) -> bool {
-    interrupt_is_sigint_only(provider, plan_keys_empty)
-        && !ready_for_input
-        && resolved_sigint_pid.is_none()
-}
-
-fn provider_turn_interrupt_plan(provider: &ProviderKind) -> Option<ProviderTurnInterruptPlan> {
-    match provider {
-        // Claude runs as a child of `agentdesk tmux-wrapper`, with stdin
-        // *piped* from the wrapper rather than wired to the PTY. A
-        // `tmux send-keys C-c` on the pane therefore delivers SIGINT to the
-        // wrapper (the PTY foreground), not to claude — and the wrapper has
-        // no SIGINT handler, so it dies and tears the pane down with it
-        // (#1260). We send SIGINT directly to claude's PID via the fallback
-        // path instead; the empty key list signals "skip send-keys, go
-        // straight to the SIGINT fallback".
-        ProviderKind::Claude => Some(ProviderTurnInterruptPlan { keys: &[] }),
-        ProviderKind::Codex => Some(ProviderTurnInterruptPlan { keys: &["Escape"] }),
-        ProviderKind::Qwen => Some(ProviderTurnInterruptPlan { keys: &["C-c"] }),
-        ProviderKind::Gemini | ProviderKind::OpenCode | ProviderKind::Unsupported(_) => None,
-    }
-}
-
-/// #3207 (part 1): how claude's TURN is cancelled without killing the session.
-/// claude has no SIGINT-safe interrupt — a direct SIGINT to the CLI exits it,
-/// the pane collapses, and the watcher tears the whole tmux session down
-/// ("dead session after turn"). The session-preserving cancel depends on how
-/// claude is hosted in the pane:
-///   * `TuiEscape` — claude runs as the interactive TUI in the pane foreground
-///     (the default `TuiHosting` driver since #2110). `tmux send-keys Escape`
-///     reaches the TUI and cancels the active generation exactly like a user
-///     pressing ESC, leaving the session alive for the next turn.
-///   * `StreamJsonControlRequest` — claude runs as a child of
-///     `agentdesk tmux-wrapper` in `--print --input-format stream-json
-///     --output-format stream-json` mode. A `control_request{subtype:interrupt}`
-///     line written to the wrapper's input FIFO is forwarded verbatim to claude
-///     stdin; the CLI acks it (`control_response{success}`), aborts the turn
-///     (`terminal_reason=aborted_streaming`), and keeps the session open for the
-///     next `user` envelope. Empirically verified against claude CLI 2.1.168.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum ClaudeTurnInterruptDelivery {
-    TuiEscape,
-    StreamJsonControlRequest,
-}
-
-/// #3207 (part 1): select the session-preserving claude interrupt mechanism from
-/// whether the tmux pane foreground is the `agentdesk tmux-wrapper` (stream-json
-/// host) or the bare claude TUI. When the pane command cannot be classified we
-/// default to `TuiEscape`: ESC is the live `TuiHosting` path, and an ESC keystroke
-/// delivered to a wrapper PTY is an inert partial-line that cannot kill anything
-/// (the wrapper only acts on complete newline-terminated JSON envelopes).
-pub(crate) fn claude_turn_interrupt_delivery(
-    pane_foreground_is_wrapper: bool,
-) -> ClaudeTurnInterruptDelivery {
-    if pane_foreground_is_wrapper {
-        ClaudeTurnInterruptDelivery::StreamJsonControlRequest
-    } else {
-        ClaudeTurnInterruptDelivery::TuiEscape
-    }
-}
-
-/// #3207 (part 1): build the stream-json interrupt control envelope the Agent
-/// SDK `interrupt()` uses. Written as a single newline-terminated line to the
-/// wrapper input FIFO, which forwards it verbatim to claude stdin.
-pub(crate) fn build_claude_interrupt_control_line(request_id: &str) -> String {
-    serde_json::json!({
-        "type": "control_request",
-        "request_id": request_id,
-        "request": { "subtype": "interrupt" }
-    })
-    .to_string()
-}
-
-fn fallback_sigint_pid_for_provider(
-    provider: &ProviderKind,
-    ready_for_input: bool,
-    provider_pid: Option<u32>,
-) -> Option<u32> {
-    match provider {
-        // #3021: Claude has NO send-keys interrupt path (empty key list — see
-        // `provider_turn_interrupt_plan`), so the direct SIGINT is both its
-        // only interrupt AND, on an idle pane, a process-kill that terminates
-        // the TUI, kills the pane, and makes the watcher tear down the whole
-        // tmux session as "dead after turn". Under `PreserveSession` (e.g. ⏳
-        // reaction removal on a finished-but-still-"active" turn) that destroys
-        // the session + context — the opposite of the policy intent. When the
-        // pane is confirmed idle (`ready_for_input`, double-checked by the
-        // caller's confirmation re-probe so a stale post-submit read cannot
-        // pass) there is no generation to interrupt, so skip the SIGINT and
-        // leave the live process alone. An actively streaming turn reports
-        // `ready_for_input == false` and is still interrupted (#1260). The
-        // hard-stop path already treats `ready_for_input` as "do not kill"
-        // (`hard_stop_pid_for_unresponsive_provider`); this mirrors it.
-        ProviderKind::Claude => {
-            if ready_for_input {
-                None
-            } else {
-                provider_pid
-            }
-        }
-        // Codex/Qwen also send Escape/C-c, but those keys reach the wrapper
-        // PTY rather than the separately-spawned provider child, so the direct
-        // SIGINT fallback is what actually stops them. The readiness probe can
-        // read a stale terminal/ready state in the sub-second window after a
-        // follow-up submit, and (unlike Claude) these providers get no
-        // confirmation re-probe — so do NOT gate their interrupt on
-        // `ready_for_input`. Always deliver the fallback when we have the child
-        // PID, matching base-branch behavior (#1260).
-        ProviderKind::Codex | ProviderKind::Qwen => provider_pid,
-        ProviderKind::Gemini | ProviderKind::OpenCode | ProviderKind::Unsupported(_) => None,
-    }
-}
-
-/// #3169 (death #3 — warm-followup teardown SIGINT self-collision): the
-/// sentinel `reason` the turn_bridge cancel epilogue records when the
-/// cancellation carried NO user-attributable `cancel_source`
-/// (`CancelToken::cancel_source()` was `None`). This is the signature of an
-/// anonymous / internal `PreserveSession` teardown — supersede, redundant
-/// delivery, or a failed watcher handoff — NOT a user-initiated stop. Every
-/// user stop path (`/stop`, `!stop`, `⏳` reaction removal, restart, skill
-/// stop) and the watchdog timeout reach `stop_active_turn` under their own
-/// descriptive reason and never produce this sentinel. Defined here and
-/// referenced by the epilogue (`turn_bridge/mod.rs`) so the producer and this
-/// SIGINT guard share a single source of truth.
-pub(in crate::services::discord) const ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON: &str =
-    "turn_bridge_cancelled";
-
-/// #3169 (death #3): claude's ONLY interrupt is a direct SIGINT (it has no
-/// send-keys path — see `provider_turn_interrupt_plan`), and on a *busy* TUI
-/// that SIGINT is a process kill (#1260; the wrapper has no SIGINT handler, so
-/// claude exits and the pane/session collapses — see
-/// `fallback_sigint_pid_for_provider`). Killing the session is the correct,
-/// desired outcome for an explicit user stop. But for an anonymous internal
-/// teardown whose whole intent is to PRESERVE the reusable session, that SIGINT
-/// destroys the just-started warm-followup turn and the entire claude session
-/// (the #3169 self-collision: a busy-queue follow-up is injected, claude starts
-/// a fresh generation, the watcher handoff for that turn fails on the
-/// #2161/#2293 quiescence-gate timeout, the same bridge turn cancels, and the
-/// teardown SIGINTs the live busy claude → exit 0 → session death).
-///
-/// Suppress the teardown SIGINT for claude on that anonymous path only. The
-/// live turn is left running for the watcher to reconcile/finalize on its next
-/// pass (the quiescence-gate already hands a stalled handoff to a
-/// deadline-armed reconciler). A genuine claude *hang* is handled by the
-/// separate stall-watchdog, so the teardown SIGINT is not the safety net for
-/// hangs and dropping it here does not strand a stuck turn. User-explicit stops
-/// keep their SIGINT (stop still works), and providers with a real send-keys
-/// interrupt (codex/qwen) are unaffected — they are never claude and their
-/// interrupt does not kill the session.
-fn claude_teardown_sigint_suppressed(provider: &ProviderKind, reason: &str) -> bool {
-    matches!(provider, ProviderKind::Claude) && reason == ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON
-}
 
 fn tmux_ready_for_input_without_tui_pane(tmux_session_name: &str, provider: &ProviderKind) -> bool {
     let binding =
@@ -1058,398 +885,6 @@ pub(in crate::services::discord) fn cancel_active_token(
 }
 
 #[cfg(unix)]
-#[derive(Debug, Clone, Eq, PartialEq)]
-struct ProcessRow {
-    pid: u32,
-    ppid: u32,
-    command: String,
-}
-
-#[cfg(unix)]
-fn provider_cli_pid_in_tmux(
-    tmux_session_name: &str,
-    provider: &ProviderKind,
-    tracked_child_pid: Option<u32>,
-) -> Option<u32> {
-    let pane_pid = crate::services::platform::tmux::pane_pid(tmux_session_name)?;
-    let rows = process_table();
-    select_provider_pid_in_pane(pane_pid, &rows, provider, tracked_child_pid)
-}
-
-/// #3207 (part 1): is the tmux pane foreground the `agentdesk tmux-wrapper`
-/// (stream-json host) rather than the bare claude TUI? Drives
-/// [`claude_turn_interrupt_delivery`]. Returns `false` when the pane / process
-/// table cannot be read, defaulting the caller to the live `TuiEscape` path.
-#[cfg(unix)]
-fn pane_foreground_is_provider_wrapper(tmux_session_name: &str) -> bool {
-    let Some(pane_pid) = crate::services::platform::tmux::pane_pid(tmux_session_name) else {
-        return false;
-    };
-    let rows = process_table();
-    rows.iter()
-        .find(|row| row.pid == pane_pid)
-        .map(|row| command_is_agentdesk_provider_wrapper(&row.command))
-        .unwrap_or(false)
-}
-
-#[cfg(not(unix))]
-fn pane_foreground_is_provider_wrapper(_tmux_session_name: &str) -> bool {
-    false
-}
-
-/// #3207 (part 1): write a pre-formatted (newline-terminated) line to the
-/// wrapper input FIFO without blocking. The wrapper holds the FIFO open
-/// `O_RDWR`, so a writer never blocks while it is alive; if the wrapper is gone
-/// the non-blocking open fails with `ENXIO` and we report the miss instead of
-/// hanging a blocking-pool thread forever.
-#[cfg(unix)]
-fn write_line_to_wrapper_fifo(input_fifo_path: &str, line: &str) -> Result<(), String> {
-    use std::io::Write;
-    use std::os::unix::fs::OpenOptionsExt;
-    let mut file = std::fs::OpenOptions::new()
-        .write(true)
-        .custom_flags(libc::O_NONBLOCK)
-        .open(input_fifo_path)
-        .map_err(|error| format!("open input fifo {input_fifo_path}: {error}"))?;
-    file.write_all(line.as_bytes())
-        .and_then(|()| {
-            if line.ends_with('\n') {
-                Ok(())
-            } else {
-                file.write_all(b"\n")
-            }
-        })
-        .and_then(|()| file.flush())
-        .map_err(|error| format!("write input fifo {input_fifo_path}: {error}"))
-}
-
-#[cfg(not(unix))]
-fn write_line_to_wrapper_fifo(_input_fifo_path: &str, _line: &str) -> Result<(), String> {
-    Err("wrapper FIFO interrupt is only supported on Unix".to_string())
-}
-
-// TUI mode regression: when the provider CLI itself is the tmux pane
-// foreground (no wrapper in front — e.g. `claude --session-id …` running
-// directly), `descendant_processes` excludes `pane_pid` and the search
-// falls through to None. The interrupt path then silently no-ops: stop
-// emoji marks the turn [Stopped] in the mailbox but no SIGINT ever
-// reaches the claude TUI, so it keeps generating and the watcher keeps
-// posting the response to Discord. Check `pane_pid` itself before walking
-// descendants.
-#[cfg(unix)]
-fn select_provider_pid_in_pane(
-    pane_pid: u32,
-    rows: &[ProcessRow],
-    provider: &ProviderKind,
-    tracked_child_pid: Option<u32>,
-) -> Option<u32> {
-    if let Some(pane_row) = rows.iter().find(|row| row.pid == pane_pid)
-        && !command_is_agentdesk_provider_wrapper(&pane_row.command)
-        && command_matches_provider(&pane_row.command, provider)
-    {
-        return Some(pane_pid);
-    }
-
-    let descendants = descendant_processes(pane_pid, rows);
-
-    if let Some(pid) = tracked_child_pid
-        && descendants.iter().any(|row| {
-            row.pid == pid
-                && !command_is_agentdesk_provider_wrapper(&row.command)
-                && command_matches_provider(&row.command, provider)
-        })
-    {
-        return Some(pid);
-    }
-
-    descendants
-        .into_iter()
-        .filter(|row| !command_is_agentdesk_provider_wrapper(&row.command))
-        .filter(|row| command_matches_provider(&row.command, provider))
-        .max_by_key(|row| provider_command_match_score(&row.command, provider))
-        .map(|row| row.pid)
-}
-
-#[cfg(not(unix))]
-fn provider_cli_pid_in_tmux(
-    _tmux_session_name: &str,
-    _provider: &ProviderKind,
-    _tracked_child_pid: Option<u32>,
-) -> Option<u32> {
-    None
-}
-
-#[cfg(unix)]
-fn process_table() -> Vec<ProcessRow> {
-    let output = match std::process::Command::new("ps")
-        .args(["-axo", "pid=,ppid=,command="])
-        .output()
-    {
-        Ok(output) if output.status.success() => output,
-        _ => return Vec::new(),
-    };
-
-    String::from_utf8_lossy(&output.stdout)
-        .lines()
-        .filter_map(parse_process_row)
-        .collect()
-}
-
-#[cfg(unix)]
-fn parse_process_row(line: &str) -> Option<ProcessRow> {
-    let mut parts = line.split_whitespace();
-    let pid = parts.next()?.parse::<u32>().ok()?;
-    let ppid = parts.next()?.parse::<u32>().ok()?;
-    let command = parts.collect::<Vec<_>>().join(" ");
-    if command.is_empty() {
-        return None;
-    }
-    Some(ProcessRow { pid, ppid, command })
-}
-
-#[cfg(unix)]
-fn descendant_processes(root_pid: u32, rows: &[ProcessRow]) -> Vec<ProcessRow> {
-    let mut descendants = Vec::new();
-    let mut stack = vec![root_pid];
-    while let Some(parent) = stack.pop() {
-        for row in rows.iter().filter(|row| row.ppid == parent) {
-            descendants.push(row.clone());
-            stack.push(row.pid);
-        }
-    }
-    descendants
-}
-
-#[cfg(unix)]
-fn command_matches_provider(command: &str, provider: &ProviderKind) -> bool {
-    provider_command_match_score(command, provider) > 0
-}
-
-#[cfg(unix)]
-fn provider_command_match_score(command: &str, provider: &ProviderKind) -> u8 {
-    let Some(binary) = provider_cli_binary_name(provider) else {
-        return 0;
-    };
-    let lower = command.to_ascii_lowercase();
-    let first = lower
-        .split_whitespace()
-        .next()
-        .unwrap_or("")
-        .trim_matches(|ch| ch == '\'' || ch == '"');
-    let first_basename = std::path::Path::new(first)
-        .file_name()
-        .and_then(|value| value.to_str())
-        .unwrap_or(first);
-
-    if first_basename == binary {
-        return 3;
-    }
-    if lower.contains(&format!("/{binary} "))
-        || lower.ends_with(&format!("/{binary}"))
-        || lower.contains(&format!(" {binary} "))
-    {
-        return 2;
-    }
-    if lower.contains(binary) { 1 } else { 0 }
-}
-
-#[cfg(unix)]
-fn command_is_agentdesk_provider_wrapper(command: &str) -> bool {
-    let lower = command.to_ascii_lowercase();
-    lower.contains(" codex-tmux-wrapper")
-        || lower.contains(" qwen-tmux-wrapper")
-        || lower.contains(" tmux-wrapper")
-}
-
-#[cfg(unix)]
-fn provider_cli_binary_name(provider: &ProviderKind) -> Option<&'static str> {
-    match provider {
-        ProviderKind::Claude => Some("claude"),
-        ProviderKind::Codex => Some("codex"),
-        ProviderKind::Qwen => Some("qwen"),
-        ProviderKind::Gemini | ProviderKind::OpenCode | ProviderKind::Unsupported(_) => None,
-    }
-}
-
-#[cfg(unix)]
-fn send_sigint(pid: u32) -> Result<(), String> {
-    #[allow(unsafe_code)]
-    let result = unsafe { libc::kill(pid as libc::pid_t, libc::SIGINT) };
-    if result == 0 {
-        Ok(())
-    } else {
-        Err(std::io::Error::last_os_error().to_string())
-    }
-}
-
-#[cfg(not(unix))]
-fn send_sigint(_pid: u32) -> Result<(), String> {
-    Err("SIGINT fallback is only supported on Unix".to_string())
-}
-
-/// #2426: wait until the given PID exits, with an upper bound. Uses OS-level
-/// exit signals (kqueue `EVFILT_PROC|NOTE_EXIT` on macOS, `pidfd_open` +
-/// `poll` on Linux) so we never busy-poll. When the PID does not exist at
-/// call time, returns immediately. When the OS signal API is unavailable or
-/// fails (e.g. permission, pre-5.3 kernel), falls back to a single bounded
-/// `tokio::sleep` so callers still observe a deterministic upper bound.
-///
-/// Returns `true` if the process exited within the deadline, `false` if the
-/// upper bound elapsed first.
-#[cfg(unix)]
-pub(super) async fn wait_for_pid_exit(pid: u32, deadline: Duration) -> bool {
-    if pid == 0 {
-        // libc::kill(0, ...) targets the whole process group; treat as "no
-        // PID to observe" and just honour the deadline.
-        tokio::time::sleep(deadline).await;
-        return false;
-    }
-
-    // Fast-path: process already gone (or never existed). `kill(pid, 0)`
-    // delivers no signal but tells us whether the PID is reachable.
-    #[allow(unsafe_code)]
-    let alive = unsafe { libc::kill(pid as libc::pid_t, 0) } == 0
-        || std::io::Error::last_os_error().raw_os_error() == Some(libc::EPERM);
-    if !alive {
-        return true;
-    }
-
-    let (tx, rx) = tokio::sync::oneshot::channel::<bool>();
-    let join = tokio::task::spawn_blocking(move || {
-        let exited = wait_for_pid_exit_blocking(pid, deadline);
-        let _ = tx.send(exited);
-    });
-
-    match tokio::time::timeout(deadline, rx).await {
-        Ok(Ok(exited)) => {
-            // Allow the blocking task to settle. It already finished (rx
-            // resolved), so this is a cheap join.
-            let _ = join.await;
-            exited
-        }
-        _ => {
-            // Outer guard fired or the blocking task panicked / dropped tx.
-            // Either way the upper bound has elapsed.
-            false
-        }
-    }
-}
-
-#[cfg(not(unix))]
-pub(super) async fn wait_for_pid_exit(_pid: u32, deadline: Duration) -> bool {
-    tokio::time::sleep(deadline).await;
-    false
-}
-
-#[cfg(target_os = "macos")]
-fn wait_for_pid_exit_blocking(pid: u32, deadline: Duration) -> bool {
-    // kqueue + EVFILT_PROC|NOTE_EXIT: kernel notifies us when the PID exits.
-    // We arm a single ONESHOT kevent and block in `kevent()` until either
-    // NOTE_EXIT fires or the supplied timespec deadline elapses.
-    #[allow(unsafe_code)]
-    unsafe {
-        let kq = libc::kqueue();
-        if kq < 0 {
-            return wait_for_pid_exit_kill_fallback(pid, deadline);
-        }
-        let change = libc::kevent {
-            ident: pid as libc::uintptr_t,
-            filter: libc::EVFILT_PROC,
-            flags: libc::EV_ADD | libc::EV_ONESHOT,
-            fflags: libc::NOTE_EXIT,
-            data: 0,
-            udata: std::ptr::null_mut(),
-        };
-        // Register the watch. A zero timeout means "register only, don't
-        // block". If the PID already exited or doesn't exist, `kevent` sets
-        // EV_ERROR with ESRCH and we treat that as "exited".
-        let zero = libc::timespec {
-            tv_sec: 0,
-            tv_nsec: 0,
-        };
-        let reg = libc::kevent(
-            kq,
-            &change as *const _,
-            1,
-            std::ptr::null_mut(),
-            0,
-            &zero as *const _,
-        );
-        if reg < 0 {
-            libc::close(kq);
-            return wait_for_pid_exit_kill_fallback(pid, deadline);
-        }
-
-        let mut event: libc::kevent = std::mem::zeroed();
-        let ts = libc::timespec {
-            tv_sec: deadline.as_secs() as libc::time_t,
-            tv_nsec: deadline.subsec_nanos() as libc::c_long,
-        };
-        let n = libc::kevent(
-            kq,
-            std::ptr::null(),
-            0,
-            &mut event as *mut _,
-            1,
-            &ts as *const _,
-        );
-        libc::close(kq);
-        if n < 0 {
-            return wait_for_pid_exit_kill_fallback(pid, deadline);
-        }
-        if n == 0 {
-            // Timeout — deadline elapsed without NOTE_EXIT.
-            return false;
-        }
-        // One event delivered. NOTE_EXIT fired or EV_ERROR/ESRCH (already
-        // gone). Both mean "PID is no longer running".
-        true
-    }
-}
-
-#[cfg(all(unix, target_os = "linux"))]
-fn wait_for_pid_exit_blocking(pid: u32, deadline: Duration) -> bool {
-    // pidfd_open(pid, 0) returns an fd that becomes readable when the
-    // process exits. We poll(POLLIN) it with the deadline. On kernels
-    // without pidfd_open (<5.3) the syscall returns ENOSYS and we fall
-    // back to a single-shot bounded wait.
-    #[allow(unsafe_code)]
-    unsafe {
-        let fd = libc::syscall(libc::SYS_pidfd_open, pid as libc::pid_t, 0_u32) as libc::c_int;
-        if fd < 0 {
-            return wait_for_pid_exit_kill_fallback(pid, deadline);
-        }
-        let mut pfd = libc::pollfd {
-            fd,
-            events: libc::POLLIN,
-            revents: 0,
-        };
-        let timeout_ms: libc::c_int = deadline.as_millis().try_into().unwrap_or(libc::c_int::MAX);
-        let n = libc::poll(&mut pfd as *mut _, 1, timeout_ms);
-        libc::close(fd);
-        if n < 0 {
-            return wait_for_pid_exit_kill_fallback(pid, deadline);
-        }
-        n > 0
-    }
-}
-
-#[cfg(all(unix, not(any(target_os = "macos", target_os = "linux"))))]
-fn wait_for_pid_exit_blocking(pid: u32, deadline: Duration) -> bool {
-    wait_for_pid_exit_kill_fallback(pid, deadline)
-}
-
-/// Last-resort fallback when the OS-native exit notifier is unavailable
-/// (kqueue creation failure, pre-5.3 Linux kernel without `pidfd_open`,
-/// other Unix). A single bounded sleep keeps the upper bound deterministic
-/// without introducing a polling loop.
-#[cfg(unix)]
-fn wait_for_pid_exit_kill_fallback(_pid: u32, deadline: Duration) -> bool {
-    std::thread::sleep(deadline);
-    false
-}
-
-#[cfg(unix)]
 pub(crate) fn tmux_runtime_paths(tmux_session_name: &str) -> (String, String) {
     use crate::services::tmux_common::session_temp_path;
     (
@@ -1526,275 +961,4 @@ pub(super) fn is_dcserver_restart_command(input: &str) -> bool {
     lower.contains("launchctl")
         && lower.contains("com.agentdesk.dcserver")
         && (lower.contains("kickstart") || lower.contains("bootstrap") || lower.contains("bootout"))
-}
-
-// #2426: tests for the PID-exit observation helper. These do not require the
-// `legacy-sqlite-tests` feature because they exercise only the
-// `wait_for_pid_exit` path and do not touch the SQLite test scaffolding.
-#[cfg(all(test, unix))]
-mod pid_exit_tests {
-    use super::wait_for_pid_exit;
-    use std::process::{Command, Stdio};
-    use std::time::{Duration, Instant};
-
-    /// A PID we are extremely unlikely to ever be reachable: well above the
-    /// typical max PID and never spawned by this test process. The kernel
-    /// reports ESRCH from `kill(pid, 0)` and `wait_for_pid_exit` should
-    /// short-circuit to `true` without blocking for the deadline.
-    #[tokio::test(flavor = "current_thread")]
-    async fn wait_for_pid_exit_returns_immediately_for_nonexistent_pid() {
-        let started = Instant::now();
-        let exited = wait_for_pid_exit(4_000_000, Duration::from_secs(5)).await;
-        let elapsed = started.elapsed();
-        assert!(
-            exited,
-            "nonexistent PID must be reported as already-exited (kill(pid,0) -> ESRCH)"
-        );
-        assert!(
-            elapsed < Duration::from_millis(500),
-            "nonexistent PID short-circuit must not honour the full deadline (took {elapsed:?})"
-        );
-    }
-
-    /// Spawn a real child, kill it, then call `wait_for_pid_exit`. The
-    /// kqueue/pidfd path must signal exit well before the 2s upper bound.
-    #[tokio::test(flavor = "current_thread")]
-    async fn wait_for_pid_exit_observes_child_termination() {
-        let mut child = Command::new("sleep")
-            .arg("30")
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .spawn()
-            .expect("spawn sleep");
-        let pid = child.id();
-
-        // Schedule a kill 50ms in the future so the helper actually has to
-        // wait for the exit signal rather than short-circuit via the
-        // `kill(pid,0)` fast path.
-        let killer = tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(50)).await;
-            // SIGKILL = unconditional, no handler chance.
-            #[allow(unsafe_code)]
-            unsafe {
-                libc::kill(pid as libc::pid_t, libc::SIGKILL);
-            }
-        });
-
-        let started = Instant::now();
-        let exited = wait_for_pid_exit(pid, Duration::from_secs(2)).await;
-        let elapsed = started.elapsed();
-
-        // Reap the zombie regardless of test outcome so we don't leak.
-        let _ = child.wait();
-        let _ = killer.await;
-
-        assert!(
-            exited,
-            "wait_for_pid_exit must report exit for killed child"
-        );
-        assert!(
-            elapsed < Duration::from_millis(1500),
-            "OS-level exit notification must beat the upper bound by a comfortable margin (took {elapsed:?})"
-        );
-    }
-
-    /// A still-running child must drive `wait_for_pid_exit` to the upper
-    /// bound and report `false`.
-    #[tokio::test(flavor = "current_thread")]
-    async fn wait_for_pid_exit_times_out_when_child_keeps_running() {
-        let mut child = Command::new("sleep")
-            .arg("30")
-            .stdin(Stdio::null())
-            .stdout(Stdio::null())
-            .stderr(Stdio::null())
-            .spawn()
-            .expect("spawn sleep");
-        let pid = child.id();
-
-        let started = Instant::now();
-        let exited = wait_for_pid_exit(pid, Duration::from_millis(200)).await;
-        let elapsed = started.elapsed();
-
-        // Tear down the still-running child.
-        #[allow(unsafe_code)]
-        unsafe {
-            libc::kill(pid as libc::pid_t, libc::SIGKILL);
-        }
-        let _ = child.wait();
-
-        assert!(
-            !exited,
-            "long-running child must not be reported as exited within the upper bound"
-        );
-        assert!(
-            elapsed >= Duration::from_millis(180),
-            "the upper bound must actually be honoured (took only {elapsed:?})"
-        );
-        assert!(
-            elapsed < Duration::from_secs(2),
-            "the upper bound must not be exceeded by more than scheduler jitter (took {elapsed:?})"
-        );
-    }
-}
-
-// #3029(A): the "missed SIGINT target" decision is a pure boolean and runs
-// under the default `cargo test` invocation (the main suite is gated behind
-// the `legacy-sqlite-tests` feature, which CI does not enable by default).
-#[cfg(test)]
-mod sigint_target_missing_tests {
-    use super::interrupt_sigint_target_missing;
-    use crate::services::provider::ProviderKind;
-
-    #[test]
-    fn active_claude_without_pid_is_a_missed_interrupt() {
-        // SIGINT-only path (empty keys = claude), turn actively generating
-        // (ready_for_input=false), and NO resolved PID → silent no-op that must
-        // now be flagged for escalation instead of reporting success.
-        assert!(
-            interrupt_sigint_target_missing(&ProviderKind::Claude, true, false, None),
-            "active claude with no resolvable PID must escalate (#3029 A), not silently succeed"
-        );
-    }
-
-    #[test]
-    fn active_claude_with_pid_is_not_missed() {
-        assert!(
-            !interrupt_sigint_target_missing(&ProviderKind::Claude, true, false, Some(42)),
-            "a resolved PID means the SIGINT had a target — not a miss"
-        );
-    }
-
-    #[test]
-    fn idle_claude_without_pid_is_not_missed() {
-        // #3021: an idle pane intentionally resolves to no PID and is left
-        // alone; that is NOT a missed interrupt.
-        assert!(
-            !interrupt_sigint_target_missing(&ProviderKind::Claude, true, true, None),
-            "idle claude (ready_for_input=true) is intentionally skipped, not a miss (#3021)"
-        );
-    }
-
-    #[test]
-    fn wrapped_providers_are_not_sigint_only() {
-        // Codex/Qwen have a send-keys path, so a missing fallback PID is not a
-        // SIGINT-only silent no-op (their send-keys still reaches the wrapper).
-        assert!(!interrupt_sigint_target_missing(
-            &ProviderKind::Codex,
-            false,
-            false,
-            None
-        ));
-        assert!(!interrupt_sigint_target_missing(
-            &ProviderKind::Qwen,
-            false,
-            false,
-            None
-        ));
-    }
-}
-
-// #3207 (part 1): the claude session-preserving interrupt selection + envelope
-// builder are pure and run under the default `cargo test` invocation (the main
-// suite is gated behind `legacy-sqlite-tests`, which CI does not enable).
-#[cfg(test)]
-mod claude_session_preserving_interrupt_tests {
-    use super::{
-        ClaudeTurnInterruptDelivery, build_claude_interrupt_control_line,
-        claude_turn_interrupt_delivery,
-    };
-
-    #[test]
-    fn tui_pane_uses_escape_keystroke() {
-        // The live `TuiHosting` driver: claude is the interactive pane fg, so ESC
-        // cancels the turn without killing the session.
-        assert_eq!(
-            claude_turn_interrupt_delivery(false),
-            ClaudeTurnInterruptDelivery::TuiEscape
-        );
-    }
-
-    #[test]
-    fn wrapper_pane_uses_stream_json_control_request() {
-        // Legacy wrapper host: claude reads stdin from a pipe, so the turn is
-        // cancelled via a stream-json control_request forwarded through the FIFO.
-        assert_eq!(
-            claude_turn_interrupt_delivery(true),
-            ClaudeTurnInterruptDelivery::StreamJsonControlRequest
-        );
-    }
-
-    #[test]
-    fn control_request_envelope_matches_agent_sdk_interrupt_shape() {
-        // Empirically verified against claude CLI 2.1.168: this exact envelope
-        // yields control_response{success} + terminal_reason=aborted_streaming
-        // and the session survives for the next user turn.
-        let line = build_claude_interrupt_control_line("req-123");
-        let parsed: serde_json::Value = serde_json::from_str(&line).expect("valid json line");
-        assert_eq!(parsed["type"], "control_request");
-        assert_eq!(parsed["request_id"], "req-123");
-        assert_eq!(parsed["request"]["subtype"], "interrupt");
-        assert!(!line.contains('\n'), "envelope is a single line");
-    }
-}
-
-// #3169 (death #3): the teardown-SIGINT suppression decision is a pure boolean
-// and runs under the default `cargo test` invocation (the main suite is gated
-// behind the `legacy-sqlite-tests` feature, which CI does not enable).
-#[cfg(test)]
-mod claude_teardown_sigint_tests {
-    use super::{ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON, claude_teardown_sigint_suppressed};
-    use crate::services::provider::ProviderKind;
-
-    #[test]
-    fn claude_anonymous_teardown_suppresses_sigint() {
-        // The forensic death path: an anonymous/internal PreserveSession
-        // teardown (no user cancel_source) reaches stop_active_turn under the
-        // sentinel reason. claude's SIGINT would kill the busy TUI + session, so
-        // it MUST be suppressed (#3169).
-        assert!(
-            claude_teardown_sigint_suppressed(
-                &ProviderKind::Claude,
-                ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON
-            ),
-            "claude on an anonymous turn_bridge teardown must NOT receive the session-killing SIGINT"
-        );
-    }
-
-    #[test]
-    fn claude_user_explicit_stop_keeps_sigint() {
-        // Every user-initiated stop passes its own descriptive reason directly
-        // to stop_active_turn — never the anonymous sentinel — so the stop
-        // feature is preserved: claude is still interrupted on an explicit stop.
-        for reason in [
-            "/stop",
-            "!stop",
-            "reaction remove ⏳",
-            "!skill stop",
-            "!cc stop",
-            "mailbox_cancel_active_turn",
-            "watchdog timeout",
-        ] {
-            assert!(
-                !claude_teardown_sigint_suppressed(&ProviderKind::Claude, reason),
-                "user-explicit / watchdog stop ({reason}) must still SIGINT claude (stop preserved)"
-            );
-        }
-    }
-
-    #[test]
-    fn non_claude_providers_are_unaffected() {
-        // codex/qwen have a real send-keys interrupt and their SIGINT does not
-        // kill the session, so the suppression never applies to them — even on
-        // the anonymous teardown reason. Existing behaviour is preserved.
-        for provider in [ProviderKind::Codex, ProviderKind::Qwen] {
-            assert!(
-                !claude_teardown_sigint_suppressed(
-                    &provider,
-                    ANONYMOUS_TURN_BRIDGE_TEARDOWN_REASON
-                ),
-                "non-claude provider must keep its existing teardown SIGINT behaviour"
-            );
-        }
-    }
 }
